+++ packages/block-editor/src/components/block-list/use-block-props/index.js
@@ -2,7 +2,6 @@
  * External dependencies
  */
 import classnames from 'classnames';
-import { omit } from 'lodash';
 
 /**
  * WordPress dependencies
@@ -10,19 +9,19 @@ import { omit } from 'lodash';
 import { useRef, useContext } from '@wordpress/element';
 import { __, sprintf } from '@wordpress/i18n';
 import { __unstableGetBlockProps as getBlockProps } from '@wordpress/blocks';
-import { useMergeRefs, useRefEffect } from '@wordpress/compose';
+import { useMergeRefs } from '@wordpress/compose';
 import { useSelect } from '@wordpress/data';
 
 /**
  * Internal dependencies
  */
 import useMovingAnimation from '../../use-moving-animation';
-import { SetBlockNodes } from '../';
 import { BlockListBlockContext } from '../block';
 import { useFocusFirstElement } from './use-focus-first-element';
 import { useIsHovered } from './use-is-hovered';
 import { useBlockMovingModeClassNames } from './use-block-moving-mode-class-names';
 import { useEventHandlers } from './use-event-handlers';
+import { useBlockNodes } from './use-block-nodes';
 import { store as blockEditorStore } from '../../../store';
 
 /**
@@ -44,7 +43,6 @@ import { store as blockEditorStore } from '../../../store';
 export function useBlockProps( props = {}, { __unstableIsHtml } = {} ) {
 	const fallbackRef = useRef();
 	const ref = props.ref || fallbackRef;
-	const setBlockNodes = useContext( SetBlockNodes );
 	const {
 		clientId,
 		isSelected,
@@ -60,46 +58,6 @@ export function useBlockProps( props = {}, { __unstableIsHtml } = {} ) {
 	const mode = useSelect( ( select ) => {
 		return select( blockEditorStore ).getBlockMode( clientId );
 	} );
-
-	// Provide the selected node, or the first and last nodes of a multi-
-	// selection, so it can be used to position the contextual block toolbar.
-	// We only provide what is necessary, and remove the nodes again when they
-	// are no longer selected.
-	const isNodeNeeded = useSelect(
-		( select ) => {
-			const {
-				isBlockSelected,
-				isFirstMultiSelectedBlock,
-				getLastMultiSelectedBlockClientId,
-			} = select( blockEditorStore );
-			return (
-				isBlockSelected( clientId ) ||
-				isFirstMultiSelectedBlock( clientId ) ||
-				getLastMultiSelectedBlockClientId() === clientId
-			);
-		}
-		// To do: figure out why tests are failing when dependencies are added.
-		// This data was originally retrieved with `withSelect` in `block.js`.
-		// For some reason, adding `clientId` as a dependency results in
-		// `toolbar-roving-tabindex.test.js` e2e test failures.
-	);
-	const nodesRef = useRefEffect(
-		( node ) => {
-			if ( ! isNodeNeeded ) {
-				return;
-			}
-
-			setBlockNodes( ( nodes ) => ( {
-				...nodes,
-				[ clientId ]: node,
-			} ) );
-
-			return () => {
-				setBlockNodes( ( nodes ) => omit( nodes, clientId ) );
-			};
-		},
-		[ isNodeNeeded, clientId, setBlockNodes ]
-	);
 
 	// translators: %s: Type of block (i.e. Text, Image etc)
 	const blockLabel = sprintf( __( 'Block: %s' ), blockTitle );
@@ -119,7 +77,7 @@ export function useBlockProps( props = {}, { __unstableIsHtml } = {} ) {
 	const htmlSuffix = mode === 'html' && ! __unstableIsHtml ? '-visual' : '';
 	const mergedRefs = useMergeRefs( [
 		ref,
-		nodesRef,
+		useBlockNodes( clientId ),
 		useEventHandlers( clientId ),
 		useIsHovered(),
 	] );
